Fake News Detection/requirements.txt
Flask==3.0.0
tensorflow==2.15.0
nltk==3.7.0
torch==2.1.1
transformers==4.14.0
torchvision==0.16.1
Pillow==10.1.0